+++ deployment/src/test/java/io/quarkiverse/fx/deployment/fxviews/FxViewRootResourceTest.java
@@ -1,29 +1,19 @@
 package io.quarkiverse.fx.deployment.fxviews;
 
-import static org.awaitility.Awaitility.await;
-
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
-
-import jakarta.enterprise.event.Observes;
 import jakarta.inject.Inject;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
-import io.quarkiverse.fx.FxPostStartupEvent;
-import io.quarkiverse.fx.QuarkusFxApplication;
-import io.quarkiverse.fx.deployment.FxTestConstants;
+import io.quarkiverse.fx.deployment.FxTestBase;
 import io.quarkiverse.fx.deployment.fxviews.controllers.RootResourceController;
 import io.quarkiverse.fx.views.FxViewData;
 import io.quarkiverse.fx.views.FxViewRepository;
-import io.quarkus.runtime.Quarkus;
 import io.quarkus.test.QuarkusUnitTest;
 import javafx.scene.layout.VBox;
 
-class FxViewRootResourceTest {
+class FxViewRootResourceTest extends FxTestBase {
 
     @RegisterExtension
     static final QuarkusUnitTest unitTest = new QuarkusUnitTest()
@@ -32,22 +22,15 @@ class FxViewRootResourceTest {
                 jar.addAsResource("RootResource.fxml");
             });
 
-    private static final AtomicBoolean eventObserved = new AtomicBoolean(false);
-
     @Inject
-    FxViewRepository fxViewRepository;
+    FxViewRepository viewRepository;
 
     @Test
     void test() {
 
-        // Non-blocking launch
-        CompletableFuture.runAsync(() -> Quarkus.run(QuarkusFxApplication.class));
+        this.startAndWait();
 
-        await()
-                .atMost(FxTestConstants.LAUNCH_TIMEOUT_MS, TimeUnit.MILLISECONDS)
-                .until(eventObserved::get);
-
-        FxViewData data = this.fxViewRepository.getViewData("RootResource");
+        FxViewData data = this.viewRepository.getViewData("RootResource");
         Assertions.assertNotNull(data);
 
         VBox root = data.getRootNode();
@@ -56,9 +39,5 @@ class FxViewRootResourceTest {
         RootResourceController controller = data.getController();
         Assertions.assertNotNull(data);
         Assertions.assertNotNull(controller.label);
-    }
-
-    void observeEvent(@Observes final FxPostStartupEvent event) {
-        eventObserved.set(true);
     }
 }
