+++ deployment/src/test/java/io/quarkiverse/fx/deployment/fxviews/FxViewTest.java
@@ -1,22 +1,12 @@
 package io.quarkiverse.fx.deployment.fxviews;
 
-import static org.awaitility.Awaitility.await;
-
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
-
-import jakarta.enterprise.event.Observes;
 import jakarta.inject.Inject;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
-import io.quarkiverse.fx.FxPostStartupEvent;
-import io.quarkiverse.fx.FxStartupLatch;
-import io.quarkiverse.fx.QuarkusFxApplication;
-import io.quarkiverse.fx.deployment.FxTestConstants;
+import io.quarkiverse.fx.deployment.FxTestBase;
 import io.quarkiverse.fx.deployment.fxviews.controllers.ComponentWithStyleController;
 import io.quarkiverse.fx.deployment.fxviews.controllers.SampleDialogController;
 import io.quarkiverse.fx.deployment.fxviews.controllers.SampleSceneController;
@@ -25,13 +15,12 @@ import io.quarkiverse.fx.deployment.fxviews.controllers.SampleTestController;
 import io.quarkiverse.fx.deployment.fxviews.controllers.SubSampleTestController;
 import io.quarkiverse.fx.views.FxViewData;
 import io.quarkiverse.fx.views.FxViewRepository;
-import io.quarkus.runtime.Quarkus;
 import io.quarkus.test.QuarkusUnitTest;
 import javafx.collections.ObservableList;
 import javafx.scene.layout.BorderPane;
 import javafx.stage.Stage;
 
-class FxViewTest {
+class FxViewTest extends FxTestBase {
 
     @RegisterExtension
     static final QuarkusUnitTest unitTest = new QuarkusUnitTest()
@@ -51,9 +40,6 @@ class FxViewTest {
     FxViewRepository viewRepository;
 
     @Inject
-    FxStartupLatch startupLatch;
-
-    @Inject
     SubSampleTestController subSampleTestController;
 
     @Inject
@@ -65,20 +51,10 @@ class FxViewTest {
     @Inject
     SampleSceneController sampleSceneController;
 
-    private static final AtomicBoolean eventObserved = new AtomicBoolean(false);
-
     @Test
-    void testFxView() throws InterruptedException {
+    void testFxView() {
 
-        // Non-blocking launch
-        CompletableFuture.runAsync(() -> Quarkus.run(QuarkusFxApplication.class));
-
-        // Wait for readiness
-        this.startupLatch.await();
-
-        await()
-                .atMost(FxTestConstants.LAUNCH_TIMEOUT_MS, TimeUnit.MILLISECONDS)
-                .until(eventObserved::get);
+        this.startAndWait();
 
         Stage primaryStage = this.viewRepository.getPrimaryStage();
         Assertions.assertNotNull(primaryStage);
@@ -105,9 +81,5 @@ class FxViewTest {
         BorderPane pane = componentWithStyle.getRootNode();
         ObservableList<String> componentStylesheets = pane.getStylesheets();
         Assertions.assertEquals(1, componentStylesheets.size());
-    }
-
-    void observeEvent(@Observes final FxPostStartupEvent event) {
-        eventObserved.set(true);
     }
 }
